Define set of default extensions.
This helps limit the amount of LANGUAGE pragmas at the top of every
module. Only small extensions are enabled by default. More intrusive
extensions, like GADT's, are enabled in the relevant files only, as
a form of documentation of which modules need these extensions.

These extra extensions are nonetheless listed in the other-extensions
field in H.cabal.