Split SEXPTYPE into its own module in order to use hsc2hs instead of c2hs for it.
We want to avoid duplicating literals from the definition of SEXPTYPE
in Rinternals.h, for forward compatibility. c2hs provides the #enum
hook, but it's not flexible enough for our needs. We want to be able
to define SEXPTYPE manually, but to refer to the literals in
Rinternals by name.