Drop hiding the Ptr in a SEXP behind a newtype.
This should allow to define a 'Storable' instance for 'HExp'. It also
makes sense because we are using storable vectors and these need
a Storable instance, i.e. for HExp too since R sometimes put SEXP's in
vectors.

Note the addition on an .hs-boot file for H.HExp, since Foreign.R and
H.HExp are now mutually recursive modules.

Also note that we dropped Data.Some, because now that SEXP is a type
synonym, it must be fully applied, and so we can no longer use
a generic higher kinded datatype such as 'Some'. Hence the
introduction of 'SomeSEXP'.