Specialize the type of ExtPtr a bit.
Switch from @Ptr a@ to @Ptr ()@. @a@ is existentially quantified so we
don't lose any expressivety, only a tiny bit of safety. But using @Ptr
()@ for pointers to values of undetermined type is fairly standard
anyways.

The reason for doing this is that it makes defining some Lift
instances less finicky (no risk of overlapping instance for Ptr).