Introduce 'R.Show' class in Prelude.
The point of introducing this class is so that 'print' can work on
'SomeSEXP' as well as 'SEXP'. Also, now we have a way to serialize
a value without side effects.

The idea is that the R prelude replicates part of the Haskell prelude,
with the same names. The 'R.Show' class has to be a different to the
Haskell 'Show' class, because R actually has two ways of serializing
a string: using deparse*() and using Rf_PrintValue(). The two produce
output that is subtly different. We aim to make the output of 'print'
be the same as R's Rf_PrintValue(), rather than use 'show' as normally
does.