Wrap SEXP's of unknown form with SomeSEXP.
Previously, many functions, such as 'eval', let the caller choose the
type of the result SEXP. This was very dangerous, because it meant the
caller could specialize the type to indicate a SEXP of completely
different form than expected. This could lead to a segfault or worse,
pattern matches on the result SEXP dispatching to the wrong branch.

The new approach is safer. If you wish to use a SEXP at a specific
type, then you have to unwrap it first, and then refine the type
through pattern matching on values produced by the 'hexp' view
function.

Another benefit is that we now have sounder typing for vector
elements, rather than fiddling with making all other R types a subtype
of ANY.