A safer approach to ensuring initialization of R, modelled after the ST monad.
Previously, we passed around an opaque 'Context' object as a witness
of R initialization. It actually wasn't much of a witness. An
unimportant issue was that the user could always cheat by passing
'undefined'. A much more important issue was that owning a 'Context'
object only proved that R had been initialized at least once at some
point in the past, /not that it is still initialized today/. Indeed,
R could have been finalized in the interim.

The new interface confers a rank-2 type to 'runR'. So instead of
having an opaque 'Context' object that we pass around, we now have an
abstract type variable that cannot escape its scope. A side benefit of
this approach is a slight performance improvement: the R monad is now
simply the (untransformed) IO monad.

To guarantee safety, we need to hide 'initialize' and especially
'finalize', lest they be called in the middle of an R computation
using 'io'.