Major overhaul of vector type indices.
Vector values used to have a type of the form `SEXP (R.Vector a)`. For
one, this was verbose, because it meant one would have to write `SEXP
(R.Vector Word8)` instead of `SEXP R.Char`. Also, it was rather
inconsistent: we were using a type index of kind `SEXPTYPE` to index
`SEXP`, but a regular Haskell type of kind `*` to index `R.Vector`.
Lastly, this scheme meant having to remember what the representation
type was used for the elements (e.g. `Word8` for `R.Char`) of a vector
of a given type, and it was easy to get it wrong.

So instead of using R.Vector as an index for all vectors, we now use
the R types directly. So `SEXP R.Char` is the type of char vectors.