Fix compile time QQ.
This commit introduces another way to use external variables which
avoids accidental capture.

foo f = [| let f = $(dyn "f") in f |]
produces
foo f = let f' = f' in f'

This is not what we want in the code that splices functions in the
Language.R.QQ. So lookupValueName is used instead.