Use preserve object to store RVals and in vector convertion.
This commit uses preserveObject and releaseObject for long lived
protection. This is better then using protect/unprotect because
objects stored in preserve object are not checked for the stack
balance.
Also this starting from this commit we are no longer use
Foreign.Concurrent.newForeignPtr so it doesn't run a deallocation
function in the separate thread.