Improve MonadR type class.
Introduce a methods that allow to create a region based approach
to R memory control. New typeclass includes following:

  * Region - associated type synonym that describes the
             current region;

  * acquire - protect and register resource inside current region.

Monad have default method and type implementations that allow to
create an instance for any monad that has MonadIO instance.

acquireSome helper added, acquireSome works like acquire but for
SomeSEXP.

R Monad was implemented in new terms.

Changes:
  1. acquire is implemented for protection of values.
  2. runRegion - is a public runner that evaluates an R monad to IO, this
       action is guaranteed to be run in R thread (that is required, but
       was not a case before)
  3. unsafeRToIO - evaluates a region but allow to pass concrete R action
       to it. This method is intended to be used internally, as some
       parts of the code require it.
  4. withEmbeddedR - now takes an IO action, that allow to pass to R thread
       only the parts of the computation that we are really interested in.

Notes:

Current approach seems to be fairly limited in sense that for any new type
that is a wrapper over SEXP we need to introduce an additional protection
method. But currently we don't have many objects of that type, all object
that we have are:

  * SEXP vectors - that have their own protection mechanisms

  * SEXP - are covered by acquire method

  * SomeSEXP - are covered by acquireSome method

There is another possible approach introduction of a typeclass:

class MonadR m => Protect m a where
  type ProtectElt m a :: *
  protect :: a -> m (ProtectElt m a)

With a sane defaults. However at the current stage this solution looks like
an over-complication.