Introduce an IsAncestor relation
IsAncestor relation introduce an 'ordering' between regions, if
'a `IsAncestor` b' holds, then all values from region a, can be
used (cast) to region b. This type family is useful even in absence
of nested regions, because even without nesting we have 3 different
types of region: V, G, s. With the ordering: forall s: V < s < G.
This addition allows to introduce a generic 'release' function, so
actions that takes a 'SEXP' from a different regions may be run in
a region rather that 'V'.

It should be noted that this API doesn't introduce any additional
complexity over previous solution, as for all cases where non generic
'release' can be used 'IsAncestor' will be automatically derived,
plus this method can be used in places where we need to loosen region
from the 'G' to 's', thats is safe but old API required to use
'unsafeRelease' that is a temporary solution and should be removed
in the final version.