Rework HExp API.
There are few major changes in HExp API:

  1. Method 'unhexp' is now a monadic action and performs
     a protection.

  2. An 'unhexpIO' method is introduced.
      UnhexpIO runs effectively runs a subregion and returnes
      an unprotected HExp value, this function has exactly the
      same properties as previous version of 'unhexp'
      (`evaluate . unhexp` to be precise) and it's possible to
      use it in quasi-quotes parser without introducing any changes.

Note:
Note, it should be carefully checked that vectors really behaves
like they should. It's definatelly safe to use vectors that got
from 'hexp' view, but it's not obvious if vectors that are got from
'mkSEXPIO' and using vectors API are also safe. The reason is that
vector allocation procedure (both pure and impure) uses protect
method under the hood, but didn't register itself in the region.