Implement quasi-quoter that works with new code.
A regional QuasiQuoter is implemented in low level terms, basically
because it leads to a very small number of changes in source code,
and because previous approach was proved to be effective and correct.

In order to do it QQ uses only IO functions that doesn't do any protection
implicitly. Another aporach may be - implement quasi quoter as a big QQ
region that will be evaluated by 'runRegion' runner (not implemented yet)
and then protect resulting value. This approach may be more safe, but
will require a big complication of QQ procedure.

In order to implement this approach following changes were made:

  * Internal function 'evalIO' was exported from the Language.R module

  * An IO version of withProtected function were introduced

       Currently we are using 'withProtected' helper that can be
       instantiated to IO monad. But with introducing a regions
       'withProtected' changed it's behaviour and can work either
       in R monad or in IO (but only inside interpreter).

       A new version of 'withProtected' work in IO and doesn't
       introduce any region based protection. This function is
       useful in lowlevel code like quasi-quoter.