Remove erroneous level of indirection in HExp that was used to indicated NULLable fields.
Some fields had type `Maybe (SEXP s a)` if from the R documentation we
inferred that this field could be NULL. However, when the
R documentation speaks of NULL, it actually means a global constant SEXP
called NilValue, not a NULL pointer. Therefore, a NULL SEXP is actually
a SEXP like any other, of type NILSXP. It is therefore wrong to map
NilValue to Nothing and all other SEXP's `s` to `Just s`, because in
doing so we break canonicity of representations: R' NilValue can now be
represented as `Nothing` or `Just Nil` (see e.g. the now retired
`heqMaybe`).

However, it is still useful to have the compiler tell us when a field is
nullable in the sense of R. Therefore, we have reinstated HExp
constraints in this same patch. These specify exactly what forms of
SEXP's are legal in each field, including whether the NILSXP form is
legal.