Fix incorrect slice operation.
This patch fixes incorrect slice operation for vectors, previously
this operation had precalculated (and wrong) offset for length.
This patch uses primitives that are provided by c2hs.

However, R doesn't support vector slicing as it required by
vectors package, that leads to segfaults. As a result it's not
safe to use operations that can't predict size of the vector.
This patch includes a fix for fromList function, however
API should still be reviewed in order to forbid user from doing
unsafe things that may lead to a memory problems.