Embed H.ghci at compile time
This makes for a rellocatable binary.

`withSystemTempFile` is used for a simplest solution: trying to pipe it
into the spawned process' stdin was futile: if we're `Inherit`ing a
pipe, we don't get a handle from `createProcess`. Attempts at `stdin`
handle failed with handle not being open to writes: perhaps we were
trying to write too early. Instead of trying to time it better, we just
use a temporary file. One caveat is that we have to close the handle
after writing otherwise the created process can't read it.