Install the necessary R packages first
We install the `numDeriv` and `optimx` packages into a temporary
directory. It does mean they get reinstalled each time but it means that
the example will work no matter where it's put (such as on read-only
partition), the global R database remains as it was and the space is
recovered after running. As it's just an example, it's very unlikely
that anyone will be running it so much that re-installs become annoying.

Regards PackageImports: to prevent clash with temporary-rc package which
stack seems to have installed. Perhaps needs investigating though it's
likely a transitive dependency somewhere.