Dummy-package nls examples
Primarily intended to explicitly specify the dependency on `mwc-random`.
An alternative approach would be to just feed in the script for the user
though that poses a question of *how* to do it: do we just assume we can
run ‘stack exec H’? It seems to defeat the point of the example.