Use temporary-rc for our temporary library
This resolves an ambiguous package import. The issue is that
`temporary-rc` is a `temporary` fork. Some time ago (1 year+),
`temporary-rc` contained extra patches and `tasty` library moved to
using that. So now we have `temporary-rc` used by `tasty` and
`temporary` we tried to use – a clash. Now the packages are functionally
the same so we use `temporary-rc`. If in the future `tasty` decides to
move back to `temporary` (as prompted at
https://github.com/batterseapower/temporary/pull/12 and @qnikst is going
to attempt to find out if there is such a plan at all), we will have to
move to `temporary`. But for now we use `temporary-rc` for simplicity.