Rewrite R quasiquoter.
The new quasiquoter follows an idea from @valderman (Anton Ekblad) for
expanding quasiquotations for higher-order languages: desugar them into
a function F in the object language, which closes over all antiquotation
variables. This approach results in much faster compile times compared
to what we were doing previously: to reconstruct at runtime the R AST
that was observed at quasiquote expansion time (i.e. compile time).

So we expand the quasiquotation into a specially crafted string,
representing F, to be parsed at runtime. This in the end isn't any
slower than evaluating code that reconstructs the AST without parsing,
because parsing the representation of F can be memoized via lazy
evaluation.

A no less important side benefit of this PR, is that we can now handle
antiquotations of higher-order values in any context, not just in
a function application context. This makes antiquoting functions into
R code much more user friendly, and less confusing. Concretely, the
following now works:

```
[r| f_hs(1) |]  -- worked previously
[r| f <- f_hs; f(1) |] -- did not work previously
[r| lapply(c(1,2,3), function(x)f_hs(x)) |] -- worked previously
[r| lapply(c(1,2,3), f_hs) |] -- did not work previously
```

One consequence of this patch is that assignments no longer have global
scope. This is a feature: it protects against effects of R's late
binding in pure quasiquotes. To get the same effect as before one has to
use the `<<-` operator. Which unlike `<-`, will conveniently fail if you
try to redefine a "locked" binding, which applies to all functions
defined in a package somewhere.