toPairList/fromPairList.
The equivalent of fromSEXP/mkSEXP for association lists / pairlists. We
can't use the Literal class for this because we target a SomeSEXP rather
than a SEXP. We can't target a SEXP because we don't know a priori the
form of the result: it could be the nil form or a cons cell.