Fix memory tests. They now pass w/ and wo/ --enable-strict-barrier.
One GC test in particular was actually incorrect, but this was never
caught because without --enable-strict-barrier, we can't properly
observe deallocation events.

So this patch further more *disables* all memory tests entirely if we
don't have --enable-strict-barrier. Since we can't reliably test
deallocation events without it. Since we can't test at runtime
whether --enable-strict-barrier was turned on or not, we instead allow
the user to pass --torture as an argument to the test suite.