Rationalize H.Prelude exports and adapt H.ghci accordingly.
As part of this patch, remove the 'Show' class - or rather, make it
private and rename it. It was being annoying before, because its name
would clash with the standard type class from the Prelude. Now the
H REPL only has H.Prelude imported by default. So the prompt is no
longer cabalistic.