Serialize execution of expQQ.
A lock for the quasiquoter was introduced in case the compiler dealt
concurrently with quasiquotations. However, the theoretical fix to the
hypothetical problem was not good since expressions returned by `parse`
do not remain protected, and if another call to `parse` happens while
the result of a former call wasn't used, then the result could be
garbage collected prematurely.

This patch makes sures no concurrent call to `parse` can happen before
the result of `parse` is used.