Drop tests in H package.
Tests in H package are not useful, because they do not add
additional test coverage over tests that are in inline-r.
Also those tests are quite hard to automate for different
workflows, as we need to expose correct working directory,
executable name and ghc package directory and it's not
easy to automate withing test itself.