Check types of incommit R values in functions.
When we wrap haskell function inside R calls we have to
test types dynamically, otherwise that may lead to segfaults
in case if value is not of expected type.

Example of a bad behavior can be found in bug #284.
After the fix function exits with error message as it should.