Fix potential problem with asynchronous exception.
There is a ponetial problem in preserve objects, when
asynchronous exception arrives on ForeignPtr allocation.
In that case resource will never be freed by the R,
leading to the memory leaks. We mask_ asynchronous exceptions
now exceptions can't arrive in the allocation block.

Thanks to @cblp (Yuriy Syrovetskiy) for noticing the
potential problem in inline-r related blogpost!